Make teuthology.lock a subpackage
