+++ teuthology/lock/__init__.py
@@ -1,0 +1,1 @@
+from teuthology.lock import cli, keys, ops, query, util  # noqa
