+++ teuthology/lock/keys.py
@@ -1,0 +1,27 @@
+import logging
+
+from teuthology import misc
+
+from . import ops, query
+
+log = logging.getLogger(__name__)
+
+
+def do_update_keys(machines, all_=False):
+    reference = query.list_locks(keyed_by_name=True)
+    if all_:
+        machines = reference.keys()
+    keys_dict = misc.ssh_keyscan(machines)
+    return push_new_keys(keys_dict, reference)
+
+
+def push_new_keys(keys_dict, reference):
+    ret = 0
+    for hostname, pubkey in keys_dict.iteritems():
+        log.info('Checking %s', hostname)
+        if reference[hostname]['ssh_pub_key'] != pubkey:
+            log.info('New key found. Updating...')
+            if not ops.update_lock(hostname, ssh_pub_key=pubkey):
+                log.error('failed to update %s!', hostname)
+                ret = 1
+    return ret
