+++ teuthology/lock/test/test_lock.py
@@ -1,8 +1,7 @@
-from teuthology import lock
-
+import teuthology.lock.util
 
 class TestLock(object):
 
     def test_locked_since_seconds(self):
         node = { "locked_since": "2013-02-07 19:33:55.000000" }
-        assert lock.locked_since_seconds(node) > 3600
+        assert teuthology.lock.util.locked_since_seconds(node) > 3600
