+++ teuthology/provision/test/test_downburst.py
@@ -21,7 +21,7 @@ class TestDownburst(object):
         ctx = self.ctx
         status = self.status
 
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         dbrst.executable = '/fake/path'
         dbrst.build_config = MagicMock(name='build_config')
@@ -43,7 +43,7 @@ class TestDownburst(object):
         ctx = self.ctx
         status = self.status
 
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         dbrst.destroy = MagicMock(name='destroy')
         dbrst.destroy.return_value = True
@@ -59,7 +59,7 @@ class TestDownburst(object):
         status = self.status
         status['locked_by'] = 'user@a'
 
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         dbrst.destroy = MagicMock(name='destroy', side_effect=RuntimeError)
 
@@ -73,7 +73,7 @@ class TestDownburst(object):
         status = self.status
         status['description'] = 'desc_a'
 
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         dbrst.destroy = MagicMock(name='destroy')
         dbrst.destroy = MagicMock(name='destroy', side_effect=RuntimeError)
@@ -82,24 +82,24 @@ class TestDownburst(object):
                                          _downburst=dbrst)
         assert result is False
 
-    @patch('teuthology.provision_executable')
+    @patch('teuthology.provision.downburst.downburst_executable')
     def test_create_fails_without_executable(self, m_exec):
         name = self.name
         ctx = self.ctx
         status = self.status
         m_exec.return_value = ''
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         result = dbrst.create()
         assert result is False
 
-    @patch('teuthology.provision_executable')
+    @patch('teuthology.provision.downburst.downburst_executable')
     def test_destroy_fails_without_executable(self, m_exec):
         name = self.name
         ctx = self.ctx
         status = self.status
         m_exec.return_value = ''
-        dbrst = provision.Downburst(
+        dbrst = provision.downburst.Downburst(
             name, ctx.os_type, ctx.os_version, status)
         result = dbrst.destroy()
         assert result is False
