+++ teuthology/suite/test/test_util.py
@@ -87,17 +87,17 @@ class TestUtil(object):
         )
         assert result == "some json"
 
-    @patch('teuthology.suite.util.lock')
-    def test_get_arch_fail(self, m_lock):
-        m_lock.list_locks.return_value = False
+    @patch('teuthology.lock.query')
+    def test_get_arch_fail(self, m_query):
+        m_query.list_locks.return_value = False
         util.get_arch('magna')
-        m_lock.list_locks.assert_called_with(machine_type="magna", count=1)
+        m_query.list_locks.assert_called_with(machine_type="magna", count=1)
 
-    @patch('teuthology.suite.util.lock')
-    def test_get_arch_success(self, m_lock):
-        m_lock.list_locks.return_value = [{"arch": "arch"}]
+    @patch('teuthology.lock.query')
+    def test_get_arch_success(self, m_query):
+        m_query.list_locks.return_value = [{"arch": "arch"}]
         result = util.get_arch('magna')
-        m_lock.list_locks.assert_called_with(
+        m_query.list_locks.assert_called_with(
             machine_type="magna",
             count=1
         )
